1.提交学习代码

+++ raladmin/raladmin-study/src/main/java/com/ral/admin/designpatterns/decoratorpatterns/Coconut.java
@@ -22,7 +22,7 @@ public class Coconut extends Ingredients {
     }
 
     @Override
-    public double cost() {
-        return 1.0 + beverage.cost();
+    public int cost() {
+        return 1 + beverage.cost();
     }
 }
